peekapp/requirements.txt
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.20
celery==3.1.18
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-allauth==0.20.0
django-bootstrap-form==3.2
django-bootstrap3==5.4.0
django-crispy-forms==1.4.0
django-datetime-widget==0.9.3
-e git+https://github.com/cypreess/django-getpaid.git@35dee9e1fb0437af3b41ff9bf3ebeb56f61d8196#egg=django_getpaid-master
django-social-auth==0.7.28
django-suit==0.2.13
-e git+https://github.com/django-tastypie/django-tastypie@256ebe1de9a78dfb5d4d6e938b813cf4c5c4ac1b#egg=django_tastypie-master
django-toolbelt==0.0.1
djangotoolbox==1.6.2
gunicorn==19.3.0
httplib2==0.9.1
kombu==3.0.26
MySQL-python==1.2.5
oauth2==1.5.211
oauthlib==0.7.2
Pillow==2.8.1
psycopg2==2.6.1
PyJWT==1.3.0
python-amazon-product-api==0.2.8
python-dateutil==2.4.2
python-mimeparse==0.1.4
python-openid==2.2.5
python-social-auth==0.2.10
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
static3==0.6.1
